requirements.txt
streamlit==1.41.1
requests==2.32.5
openai==1.109.1
boto3==1.39.3
qdrant-client==1.15.1
tiktoken==0.9.0
trafilatura==1.12.2
tenacity==9.1.2
python-dotenv==1.1.1
